% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_map_chart.R
\name{as_exposure_map_data}
\alias{as_exposure_map_data}
\title{Convert raw data into a exposure map data frame}
\usage{
as_exposure_map_data(
  results_map,
  start_year,
  dataframe_translations,
  language_select = "EN"
)
}
\arguments{
\item{results_map}{results_map}

\item{start_year}{start_year}

\item{dataframe_translations}{dataframe_translations}

\item{language_select}{two letter code for language (single string; default = "EN")}
}
\description{
The \code{as_exposure_map_data} function converts raw data into a exposure map data frame
}
