% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_comparison_chart.R
\name{as_peer_comparison_data}
\alias{as_peer_comparison_data}
\title{Convert raw data into a peer comparison data frame}
\usage{
as_peer_comparison_data(
  investor_name,
  portfolio_name,
  peer_group,
  start_year,
  equity_results_portfolio,
  bonds_results_portfolio,
  peers_equity_results_user,
  peers_bonds_results_user,
  select_scenario,
  select_scenario_auto,
  select_scenario_shipping,
  select_scenario_other,
  dataframe_translations,
  language_select = "EN"
)
}
\arguments{
\item{investor_name}{investor_name}

\item{portfolio_name}{portfolio_name}

\item{peer_group}{peer_group}

\item{start_year}{start_year}

\item{equity_results_portfolio}{equity_results_portfolio}

\item{bonds_results_portfolio}{bonds_results_portfolio}

\item{peers_equity_results_user}{peers_equity_results_user}

\item{peers_bonds_results_user}{peers_bonds_results_user}

\item{select_scenario}{select_scenario}

\item{select_scenario_auto}{select_scenario_auto}

\item{select_scenario_shipping}{select_scenario_shipping}

\item{select_scenario_other}{select_scenario_other}

\item{dataframe_translations}{dataframe_translations}

\item{language_select}{two letter code for language (single string; default = "EN")}
}
\description{
The \code{as_peer_comparison_data} function converts raw data into a peer comparison data frame
}
