% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tech_exposure_chart.R
\name{as_tech_exposure_data}
\alias{as_tech_exposure_data}
\title{Convert raw data into a tech exposure data frame}
\usage{
as_tech_exposure_data(
  investor_name,
  portfolio_name,
  start_year,
  peer_group,
  equity_results_portfolio,
  bonds_results_portfolio,
  indices_equity_results_portfolio,
  indices_bonds_results_portfolio,
  green_techs = c("RenewablesCap", "HydroCap", "NuclearCap", "Hybrid", "Electric",
    "FuelCell", "Hybrid_HDV", "Electric_HDV", "FuelCell_HDV", "Ac-Electric Arc Furnace",
    "Dc-Electric Arc Furnace"),
  select_scenario,
  select_scenario_auto,
  select_scenario_shipping,
  select_scenario_other,
  all_tech_levels,
  equity_market_levels,
  dataframe_translations,
  language_select = "EN"
)
}
\arguments{
\item{investor_name}{investor_name}

\item{portfolio_name}{portfolio_name}

\item{start_year}{start_year}

\item{peer_group}{peer_group}

\item{equity_results_portfolio}{equity_results_portfolio}

\item{bonds_results_portfolio}{bonds_results_portfolio}

\item{indices_equity_results_portfolio}{indices_equity_results_portfolio}

\item{indices_bonds_results_portfolio}{indices_bonds_results_portfolio}

\item{green_techs}{green_techs}

\item{select_scenario}{select_scenario}

\item{select_scenario_auto}{select_scenario_auto}

\item{select_scenario_shipping}{select_scenario_shipping}

\item{select_scenario_other}{select_scenario_other}

\item{all_tech_levels}{all_tech_levels}

\item{equity_market_levels}{equity_market_levels}

\item{dataframe_translations}{dataframe_translations}

\item{language_select}{two letter code for language (single string; default = "EN")}
}
\description{
The \code{as_tech_exposure_data} function converts raw data into a tech exposure data frame
}
