% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_alignmemt_chart.R
\name{as_trajectory_alignment_data}
\alias{as_trajectory_alignment_data}
\title{Convert raw data into a trajectory alignment data frame}
\usage{
as_trajectory_alignment_data(
  equity_results_portfolio,
  bonds_results_portfolio,
  indices_equity_results_portfolio,
  indices_bonds_results_portfolio,
  investor_name,
  portfolio_name,
  tech_roadmap_sectors,
  scen_geo_levels,
  all_tech_levels,
  year_horizon = 5,
  dataframe_translations,
  language_select = "en"
)
}
\arguments{
\item{equity_results_portfolio}{equity_results_portfolio}

\item{bonds_results_portfolio}{bonds_results_portfolio}

\item{indices_equity_results_portfolio}{indices_equity_results_portfolio}

\item{indices_bonds_results_portfolio}{indices_bonds_results_portfolio}

\item{investor_name}{investor_name}

\item{portfolio_name}{portfolio_name}

\item{tech_roadmap_sectors}{tech_roadmap_sectors}

\item{scen_geo_levels}{scen_geo_levels}

\item{all_tech_levels}{all_tech_levels}

\item{year_horizon}{number of years to include from the start year (single integer/numeric)}

\item{dataframe_translations}{dataframe_translations}

\item{language_select}{two letter code for language (single string; default = "EN")}
}
\description{
The \code{as_trajectory_alignment_data} function converts raw data into a trajectory alignment data frame
}
