% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_map_chart.R
\name{exposure_map_chart}
\alias{exposure_map_chart}
\title{Create an interactive exposure map chart in an htmlwidget}
\usage{
exposure_map_chart(.data, width = NULL, height = NULL, ...)
}
\arguments{
\item{.data}{an exposure map data frame}

\item{width}{width of exported htmlwidget in pixels (single integer value; default == NULL)}

\item{height}{height of exported htmlwidget in pixels (single integer value; default == NULL)}

\item{...}{other options passed on to r2d3::r2d3() (see details)}
}
\description{
The \code{exposure_map_chart} function creates an interactive exposure map chart in an htmlwidget
}
