% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_pie_chart.R
\name{exposure_pie_chart}
\alias{exposure_pie_chart}
\title{Create an interactive exposure pie chart in an htmlwidget}
\usage{
exposure_pie_chart(.data, width = NULL, height = NULL, ...)
}
\arguments{
\item{.data}{a exposure pie data frame}

\item{width}{width of exported htmlwidget in pixels (single integer value; default == NULL)}

\item{height}{height of exported htmlwidget in pixels (single integer value; default == NULL)}

\item{...}{other options passed on to r2d3::r2d3() (see details)}
}
\description{
The \code{exposure_pie_chart} function creates an interactive exposure pie chart in an htmlwidget
}
