% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_comparison_chart.R
\name{peer_comparison_chart}
\alias{peer_comparison_chart}
\title{Create an interactive peer comparison chart in an htmlwidget}
\usage{
peer_comparison_chart(.data, width = NULL, height = NULL, ...)
}
\arguments{
\item{.data}{a peer comparison data frame}

\item{width}{width of exported htmlwidget in pixels (single integer value; default == NULL)}

\item{height}{height of exported htmlwidget in pixels (single integer value; default == NULL)}

\item{...}{other options passed on to r2d3::r2d3() (see details)}
}
\description{
The \code{peer_comparison_chart} function creates an interactive peer comparison chart in an htmlwidget
}
