% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tech_exposure_chart.R
\name{tech_exposure_chart}
\alias{tech_exposure_chart}
\title{Create an interactive tech exposure chart in an htmlwidget}
\usage{
tech_exposure_chart(.data, width = NULL, height = NULL, ...)
}
\arguments{
\item{.data}{a tech exposure data frame}

\item{width}{width of exported htmlwidget in pixels (single integer value; default == NULL)}

\item{height}{height of exported htmlwidget in pixels (single integer value; default == NULL)}

\item{...}{other options passed on to r2d3::r2d3() (see details)}
}
\description{
The \code{tech_exposure_chart} function creates an interactive tech exposure chart in an htmlwidget
}
