% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_alignmemt_chart.R
\name{trajectory_alignment_chart}
\alias{trajectory_alignment_chart}
\title{Create an interactive trajectory alignment chart in an htmlwidget}
\usage{
trajectory_alignment_chart(.data, width = NULL, height = NULL, ...)
}
\arguments{
\item{.data}{a trajectory alignment data frame}

\item{width}{width of exported htmlwidget in pixels (single integer value; default == NULL)}

\item{height}{height of exported htmlwidget in pixels (single integer value; default == NULL)}

\item{...}{other options passed on to r2d3::r2d3() (see details)}
}
\description{
The \code{trajectory_alignment_chart} function creates an interactive trajectory alignment chart in an htmlwidget
}
