% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_pie_chart.R
\name{as_exposure_pie_data}
\alias{as_exposure_pie_data}
\title{Convert raw data into a exposure pie data frame}
\usage{
as_exposure_pie_data(
  audit_file,
  investor_name,
  portfolio_name,
  pacta_sectors = c("Power", "Automotive", "Shipping", "Oil&Gas", "Coal", "Steel",
    "Cement", "Aviation"),
  dataframe_translations,
  language_select = "EN"
)
}
\arguments{
\item{audit_file}{audit_file}

\item{investor_name}{investor_name}

\item{portfolio_name}{portfolio_name}

\item{pacta_sectors}{pacta_sectors}

\item{dataframe_translations}{dataframe_translations}

\item{language_select}{two letter code for language (single string; default = "EN")}
}
\description{
The \code{as_exposure_pie_data} function converts raw data into a exposure pie data frame
}
