% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_alignmemt_chart.R
\name{as_trajectory_alignment_data}
\alias{as_trajectory_alignment_data}
\title{Convert raw data into a trajectory alignment data frame}
\usage{
as_trajectory_alignment_data(
  investor_name,
  portfolio_name,
  peer_group,
  equity_results_portfolio,
  bonds_results_portfolio,
  indices_equity_results_portfolio,
  indices_bonds_results_portfolio,
  peers_equity_results_portfolio,
  peers_bonds_results_portfolio,
  tech_roadmap_sectors,
  scen_geo_levels,
  all_tech_levels,
  dataframe_translations,
  language_select = "EN"
)
}
\arguments{
\item{investor_name}{investor_name}

\item{portfolio_name}{portfolio_name}

\item{peer_group}{peer_group}

\item{equity_results_portfolio}{equity_results_portfolio}

\item{bonds_results_portfolio}{bonds_results_portfolio}

\item{indices_equity_results_portfolio}{indices_equity_results_portfolio}

\item{indices_bonds_results_portfolio}{indices_bonds_results_portfolio}

\item{peers_equity_results_portfolio}{peers_equity_results_portfolio}

\item{peers_bonds_results_portfolio}{peers_bonds_results_portfolio}

\item{tech_roadmap_sectors}{tech_roadmap_sectors}

\item{scen_geo_levels}{scen_geo_levels}

\item{all_tech_levels}{all_tech_levels}

\item{dataframe_translations}{dataframe_translations}

\item{language_select}{two letter code for language (single string; default = "EN")}
}
\description{
The \code{as_trajectory_alignment_data} function converts raw data into a trajectory alignment data frame
}
